PI-3778 event number as string in entity bug cast to Int

+++ projects/pre-sentence-reports-to-delius/src/main/kotlin/uk/gov/justice/digital/hmpps/service/DetailsService.kt
@@ -17,7 +17,7 @@ class DetailsService
     fun getDefendantDetails(psrUuid: String): DefendantDetails {
         val document = documentRepository.getByUuid(psrUuid)
         val person = document.person
-        val eventNumber = Integer.valueOf(document.courtReport.courtAppearance.event.eventNumber).toLong()
+        val eventNumber = Integer.valueOf(document.courtReport.courtAppearance.event.eventNumber)
         val name = person.name()
         val address = personAddressRepository.findMainAddress(person.id)?.let {
             Address(
